+++ packages/js/email-editor/src/store/types.ts
@@ -38,6 +38,7 @@ export type ExperimentalSettings = {
 export type EmailEditorSettings = EditorSettings &
 	ExperimentalSettings & {
 		isPreviewMode: boolean;
+		allowedIframeStyleHandles?: string[];
 	};
 
 export type EmailTheme = {