+++ plugins/woocommerce/client/blocks/tests/e2e/tests/stock-filter/stock-filter.block_theme.spec.ts
@@ -6,6 +6,7 @@ import {
 	expect,
 	TemplateCompiler,
 	wpCLI,
+	BLOCK_THEME_SLUG,
 } from '@woocommerce/e2e-utils';
 
 export const blockData = {
@@ -121,7 +122,7 @@ test.describe( `${ blockData.name } Block - with PHP classic template`, () => {
 		await page.reload();
 
 		await admin.visitSiteEditor( {
-			postId: 'woocommerce/woocommerce//archive-product',
+			postId: `${ BLOCK_THEME_SLUG }//archive-product`,
 			postType: 'wp_template',
 			canvas: 'edit',
 		} );